RepositoryDetailsInfoView を外部ファイル化（プロンプトが重くなりすぎる）
